Crud de libros hecho

+++ src/main/java/com/sebasPortillo/Respository/GenderRepository.java
@@ -9,5 +9,5 @@ import java.util.List;
 public interface GenderRepository extends JpaRepository<Gender,Long> {
 
     @Query(value = "select g.* from genero g, generos_libro gl where g.id = gl.fk_genero and gl.fk_libro = :idBook", nativeQuery = true)
-    public List<Gender> findGenderByBook(int idBook);
+    public List<Gender> findGenderByBook(long idBook);
 }
